requirements.txt
altgraph==0.17
future==0.18.2
pefile==2021.5.24
Pillow==8.3.1
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
pywin32-ctypes==0.2.0
altgraph==0.17
annotated-types==0.5.0
anyio==3.7.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.5
exceptiongroup==1.1.2
fastapi==0.100.0
future==0.18.2
h11==0.14.0
idna==3.4
macholib==1.16.2
pefile==2021.5.24
Pillow==8.3.1
pydantic==2.0.3
pydantic_core==2.3.0
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
pywin32-ctypes==0.2.0
requests==2.28.2
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.15
uvicorn==0.23.0
